+++ src/jmetal/metaheuristics/agmopso/AgMOPSnich.java
@@ -1,4 +1,5 @@
 package jmetal.metaheuristics.agmopso;
+
 import jmetal.core.*;
 import jmetal.util.Distance;
 import jmetal.util.JMException;
@@ -42,8 +43,8 @@ public class AgMOPSnich extends Algorithm {
 		maxIterations = ((Integer) this.getInputParameter("maxIterations")).intValue();
 		populationSize = ((Integer) this.getInputParameter("swarmSize")).intValue();
 		archive = new CrowdingArchive(populationSize, problem.getNumberOfObjectives());
-		int clonesize = (int) populationSize / 5;
-		T_ = (int) populationSize / 5;
+		int clonesize = populationSize / 5;
+		T_ = populationSize / 5;
 		SolutionSet clonepopulation = new SolutionSet(clonesize);
 		int evelations = 0;
 		int max_evelations = populationSize * maxIterations;
@@ -141,9 +142,9 @@ public class AgMOPSnich extends Algorithm {
 			for (i = 0; i <= H_; i++) {
 				for (j = 0; j <= H_; j++) {
 					if (i + j <= H_) {
-						lamdaVectors[nw][0] = (double) (1.0 * i) / H_;
-						lamdaVectors[nw][1] = (double) (1.0 * j) / H_;
-						lamdaVectors[nw][2] = (double) (1.0 * (H_ - i - j) / H_);
+						lamdaVectors[nw][0] = (1.0 * i) / H_;
+						lamdaVectors[nw][1] = (1.0 * j) / H_;
+						lamdaVectors[nw][2] = 1.0 * (H_ - i - j) / H_;
 						nw++;
 					} // if
 				} // for
@@ -157,11 +158,11 @@ public class AgMOPSnich extends Algorithm {
 					for (c = 0; c <= H_; c++) {
 						for (d = 0; d <= H_; d++) {
 							if (a + b + c + d <= H_) {
-								lamdaVectors[nw][0] = (double) (1.0 * a) / H_;
-								lamdaVectors[nw][1] = (double) (1.0 * b) / H_;
-								lamdaVectors[nw][2] = (double) (1.0 * c) / H_;
-								lamdaVectors[nw][3] = (double) (1.0 * d) / H_;
-								lamdaVectors[nw][4] = (double) (1.0 * (H_ - a - b - c - d) / H_);
+								lamdaVectors[nw][0] = (1.0 * a) / H_;
+								lamdaVectors[nw][1] = (1.0 * b) / H_;
+								lamdaVectors[nw][2] = (1.0 * c) / H_;
+								lamdaVectors[nw][3] = (1.0 * d) / H_;
+								lamdaVectors[nw][4] = 1.0 * (H_ - a - b - c - d) / H_;
 								nw++;
 							}
 						}
@@ -183,14 +184,14 @@ public class AgMOPSnich extends Algorithm {
 								for (f = 0; f <= H1_; f++) {
 									for (g = 0; g <= H1_; g++) {
 										if (a + b + c + d + e + f + g <= H1_) {
-											lambda1[nw1][0] = (double) (1.0 * a) / H1_;
-											lambda1[nw1][1] = (double) (1.0 * b) / H1_;
-											lambda1[nw1][2] = (double) (1.0 * c) / H1_;
-											lambda1[nw1][3] = (double) (1.0 * d) / H1_;
-											lambda1[nw1][4] = (double) (1.0 * e) / H1_;
-											lambda1[nw1][5] = (double) (1.0 * f) / H1_;
-											lambda1[nw1][6] = (double) (1.0 * g) / H1_;
-											lambda1[nw1][7] = (double) (1.0 * (H1_ - a - b - c - d - e - f - g) / H1_);
+											lambda1[nw1][0] = (1.0 * a) / H1_;
+											lambda1[nw1][1] = (1.0 * b) / H1_;
+											lambda1[nw1][2] = (1.0 * c) / H1_;
+											lambda1[nw1][3] = (1.0 * d) / H1_;
+											lambda1[nw1][4] = (1.0 * e) / H1_;
+											lambda1[nw1][5] = (1.0 * f) / H1_;
+											lambda1[nw1][6] = (1.0 * g) / H1_;
+											lambda1[nw1][7] = 1.0 * (H1_ - a - b - c - d - e - f - g) / H1_;
 											nw1++;
 										}
 									}
@@ -209,14 +210,14 @@ public class AgMOPSnich extends Algorithm {
 								for (f = 0; f <= H2_; f++) {
 									for (g = 0; g <= H2_; g++) {
 										if (a + b + c + d + e + f + g <= H2_) {
-											lambda2[nw2][0] = (double) (1.0 * a) / H2_;
-											lambda2[nw2][1] = (double) (1.0 * b) / H2_;
-											lambda2[nw2][2] = (double) (1.0 * c) / H2_;
-											lambda2[nw2][3] = (double) (1.0 * d) / H2_;
-											lambda2[nw2][4] = (double) (1.0 * e) / H2_;
-											lambda2[nw2][5] = (double) (1.0 * f) / H2_;
-											lambda2[nw2][6] = (double) (1.0 * g) / H2_;
-											lambda2[nw2][7] = (double) (1.0 * (H2_ - a - b - c - d - e - f - g) / H2_);
+											lambda2[nw2][0] = (1.0 * a) / H2_;
+											lambda2[nw2][1] = (1.0 * b) / H2_;
+											lambda2[nw2][2] = (1.0 * c) / H2_;
+											lambda2[nw2][3] = (1.0 * d) / H2_;
+											lambda2[nw2][4] = (1.0 * e) / H2_;
+											lambda2[nw2][5] = (1.0 * f) / H2_;
+											lambda2[nw2][6] = (1.0 * g) / H2_;
+											lambda2[nw2][7] = 1.0 * (H2_ - a - b - c - d - e - f - g) / H2_;
 											nw2++;
 										}
 									}
@@ -259,16 +260,16 @@ public class AgMOPSnich extends Algorithm {
 										for (h = 0; h <= H1_; h++) {
 											for (i = 0; i <= H1_; i++) {
 												if (a + b + c + d + e + f + g + h + i <= H1_) {
-													lambda1[nw1][0] = (double) (1.0 * a) / H1_;
-													lambda1[nw1][1] = (double) (1.0 * b) / H1_;
-													lambda1[nw1][2] = (double) (1.0 * c) / H1_;
-													lambda1[nw1][3] = (double) (1.0 * d) / H1_;
-													lambda1[nw1][4] = (double) (1.0 * e) / H1_;
-													lambda1[nw1][5] = (double) (1.0 * f) / H1_;
-													lambda1[nw1][6] = (double) (1.0 * g) / H1_;
-													lambda1[nw1][7] = (double) (1.0 * h) / H1_;
-													lambda1[nw1][8] = (double) (1.0 * i) / H1_;
-													lambda1[nw1][9] = (double) (1.0 * (H1_ - a - b - c - d - e - f - g - h - i) / H1_);
+													lambda1[nw1][0] = (1.0 * a) / H1_;
+													lambda1[nw1][1] = (1.0 * b) / H1_;
+													lambda1[nw1][2] = (1.0 * c) / H1_;
+													lambda1[nw1][3] = (1.0 * d) / H1_;
+													lambda1[nw1][4] = (1.0 * e) / H1_;
+													lambda1[nw1][5] = (1.0 * f) / H1_;
+													lambda1[nw1][6] = (1.0 * g) / H1_;
+													lambda1[nw1][7] = (1.0 * h) / H1_;
+													lambda1[nw1][8] = (1.0 * i) / H1_;
+													lambda1[nw1][9] = 1.0 * (H1_ - a - b - c - d - e - f - g - h - i) / H1_;
 													nw1++;
 												}
 											}
@@ -291,16 +292,16 @@ public class AgMOPSnich extends Algorithm {
 										for (h = 0; h <= H2_; h++) {
 											for (i = 0; i <= H2_; i++) {
 												if (a + b + c + d + e + f + g + h + i <= H2_) {
-													lambda1[nw2][0] = (double) (1.0 * a) / H2_;
-													lambda1[nw2][1] = (double) (1.0 * b) / H2_;
-													lambda1[nw2][2] = (double) (1.0 * c) / H2_;
-													lambda1[nw2][3] = (double) (1.0 * d) / H2_;
-													lambda1[nw2][4] = (double) (1.0 * e) / H2_;
-													lambda1[nw2][5] = (double) (1.0 * f) / H2_;
-													lambda1[nw2][6] = (double) (1.0 * g) / H2_;
-													lambda1[nw2][7] = (double) (1.0 * h) / H2_;
-													lambda1[nw2][8] = (double) (1.0 * i) / H2_;
-													lambda1[nw2][9] = (double) (1.0 * (H2_ - a - b - c - d - e - f - g - h - i) / H2_);
+													lambda1[nw2][0] = (1.0 * a) / H2_;
+													lambda1[nw2][1] = (1.0 * b) / H2_;
+													lambda1[nw2][2] = (1.0 * c) / H2_;
+													lambda1[nw2][3] = (1.0 * d) / H2_;
+													lambda1[nw2][4] = (1.0 * e) / H2_;
+													lambda1[nw2][5] = (1.0 * f) / H2_;
+													lambda1[nw2][6] = (1.0 * g) / H2_;
+													lambda1[nw2][7] = (1.0 * h) / H2_;
+													lambda1[nw2][8] = (1.0 * i) / H2_;
+													lambda1[nw2][9] = 1.0 * (H2_ - a - b - c - d - e - f - g - h - i) / H2_;
 													nw2++;
 												}
 											}
@@ -454,7 +455,6 @@ public class AgMOPSnich extends Algorithm {
 			if (fitnessFunction(leader_ind.get(i), lamdaVectors[i]) > minFit) {
 				double d = diversity(archive.get(best_ind), this.lamdaVectors[i]);
 				if (d > max_d) {
-					;
 				}
 				max_d = d;
 				archive.get(best_ind).setCrowdingDistance(d / max_d);
@@ -635,6 +635,7 @@ public class AgMOPSnich extends Algorithm {
 	public String getname() {
 		return "AGMOPSOnich";
 	}
+
 	double diversity(Solution individual, double[] lamda) {
 		int i;
 		double d1, nl;
